+++ _meta.lua
@@ -6,7 +6,7 @@ local _ = require("gettext")
 return {
     name = "updatesmanager",
     fullname = _("Updates Manager"),
-    description = _([[Manage updates for patches and plugins from multiple GitHub repositories. Check for updates, view patch information, and update with ease.]]),
-    version = "1.0.6",
+    description = _([[Manage updates for patches and plugins GitHub repositories.]]),
+    version = "1.0.7",
 }